Fixed some bugs in the reorder method and added a delete workoutExercise method

+++ src/main/java/com/lukeroche/fit/domain/dto/WorkoutExerciseResponse.java
@@ -15,7 +15,7 @@ public class WorkoutExerciseResponse {
 
     private Long id;
 
-    private Integer orderIndex;
+    private Long orderIndex;
 
     private String notes;
